requirements.txt
streamlit
firebase-admin
pandas
plotly
openpyxl
pyrebase4